refactor: extract common ssr example testing code
