+++ tools/ssr-e2e-harness/Cargo.toml
@@ -8,3 +8,4 @@ rust-version.workspace = true
 gloo = { workspace = true, features = ["futures"] }
 web-sys = { workspace = true, features = ["Performance", "PerformanceEntry"] }
 wasm-bindgen = { workspace = true }
+yew = { path = "../../packages/yew", features = ["csr", "hydration", "test"] }
